Fix build for windows/386 platform
Before this patch, the build on 32bit windows would fail with:

```
> GOOS=windows GOARCH=386 go build ./...
./term_windows.go:74:39: constant 4294967286 overflows int
./term_windows.go:80:40: constant 4294967285 overflows int
./term_windows.go:86:40: constant 4294967284 overflows int
```

We now cast the values to int64 to avoid this error.

+++ term_windows.go
@@ -71,19 +71,19 @@ func StdStreams() (stdIn io.ReadCloser, stdOut, stdErr io.Writer) {
 	// go-ansiterm hasn't switch to x/sys/windows.
 	// TODO: switch back to x/sys/windows once go-ansiterm has switched
 	if emulateStdin {
-		stdIn = windowsconsole.NewAnsiReader(windows.STD_INPUT_HANDLE)
+		stdIn = windowsconsole.NewAnsiReader(int64(windows.STD_INPUT_HANDLE)) // nolint: unconvert
 	} else {
 		stdIn = os.Stdin
 	}
 
 	if emulateStdout {
-		stdOut = windowsconsole.NewAnsiWriter(windows.STD_OUTPUT_HANDLE)
+		stdOut = windowsconsole.NewAnsiWriter(int64(windows.STD_OUTPUT_HANDLE)) // nolint: unconvert
 	} else {
 		stdOut = os.Stdout
 	}
 
 	if emulateStderr {
-		stdErr = windowsconsole.NewAnsiWriter(windows.STD_ERROR_HANDLE)
+		stdErr = windowsconsole.NewAnsiWriter(int64(windows.STD_ERROR_HANDLE)) // nolint: unconvert
 	} else {
 		stdErr = os.Stderr
 	}
